nbv_explore_net/requirements-no-pytorch.txt
numpy
open3d
opencv-python
scipy
tqdm
attrdict
pillow
scikit-learn 
scikit-spatial
plotly
trimesh
jupyter
pyrender
torch-geometric
torch-cluster
torch-scatter
glob2==0.5
pybullet==3.2.5